Build Cargo dynamically linked
We originally statically linked Cargo because it would make it easier to
send to developer machines were it involved in the Android tree build.
Since the current plan is to only use Cargo for building rustc, it only
needs to run on the buildbots.

This also helps make the output toolchain more reproducible.

Bug: 140823999
Change-Id: Iee4ef6bdfac35893dde71c00a127ef4f1bf96f1e