Use [patch] instead of manually updating checksums
Vendored sources usually use "directory sources", which are not tracked
for rebuilds. To avoid confusing users, rustc will hash the files in a
directory source, and error if they our modified. We carry a patch to
compiler_builtins, and may carry more to vendor directories in the
future.

Previously, I manually updated the expected hash, but this is obviously
brittle. In this update, I instead use a [patch] entry in Cargo.toml to
redirect the dependency to the same path, but as a "path" source, which
is not hashed (and will be checked for updates on rebuild).

In order to support this, we regenerate the lockfile during the build
process after configuring but before running the real build.

Bug: 152633654
Change-Id: I29daf820538bd792ce8ceca71d276f16cc6a9da6