This CL replaces two exiting patches with a new one
This CL adds a patch to be applied to rustc versions 1.53 and greater.
This patch will update sanitizer support definitions for Android inside
the compiler as well as adjusts the logic to make sure no pre-built
sanitizer runtime libraries are emitted as linktime requirements.

This patch supersedes the previous two patches.

Bug: 194400612
Change-Id: I56238a5b38da70cc4568a586313dee0d2df7feb2