Remove the Python3 shim
Prior to cl/397099309 the buildbots used the python interpreter
available in the build-tools project.  Now the buildbots will use the
prebuilt python3 interpreter that is part of the toolchain/android_rust
project.  As a result of this we can remove the shim python2 script that
would invoke the actual build script using python3.

Test: ./build.py
Change-Id: Iddd69c0d80894ef685d99171a362e00203fce595