Improved handling of branching
This CL updates fetch_source.py with some additional logic to better
handle branching.  It will now test for the existence of a branch before
attempting to create it.  If it exists and you have passed the
--overwrite flag it will check out the branch and amend it.  If the
branch exists and the flag wasn't set the script will raise an error.

Test: `./fetch_source.py -b 1.56.0` with and without the branch existing
Change-Id: I6d464e6428f837cf4e563b979c448bc6e534f119