Add escaping back to the use of `$ORIGIN`.
A refactoring accidentally dropped the shell escaping of `$ORIGIN`,
leading it to be expanded as a regular shell variable, rather than
passing `$ORIGIN/../lib64` as part of the RUNPATH for rustc and other
binaries. This results in broken binaries for anything other than the
system in which they were built (because the binaries do have an
absolute path to the libc++.so.1 from that particular toolchain). Future
work might include removing that absolute path because it can mask
issues like this, since local tests would still pass unexpectedly.

Test: `./build.py` and use `llvm-readelf` to verify `RUNPATH` of rustc
Change-Id: Ia357a43b0dfdbbb3c9d49d2b873adfd6c388edc1