Improved dynamic loading path handling
In aosp/1819023 the escape characters for the `$ORIGIN` value of the
rpath flag passed to the linker was dropped.  This resulted in (after
shell expansion) an rpath value of `/../lib64` instead of
`$ORIGIN/../lib64`, which in turn caused the executable to fail to load
when it couldn't find `libc++.so.1`.

This issue was hidden during previous testing because, while the
relative path was incorrect, the absolute path used to find libraries
during buildtime was still available on the machine that built the
artifacts.  As a result the artifacts could still locate appropriate
libraries, outside of the prebuilt path, that would not be available on
other machines.

To address this problem the buildtime dynamic library path is now
specified via LD_LIBRARY_PATH.  This prevents buildtime values from
polluting generated artifacts and will help to ensure that our builds
are more hermetic by ensuring that the produced binaries rely on
libraries that are included in the prebuilt archive.

Test: ./build.py --lto thin
Test: Check generated binaries' rpath values
Change-Id: I6e92216f73b0e5e27221416ddaa5ad0d4e1ba4f5