Fix a broken patch command (on darwin).
A `-` parameter to the `-r` option was dropped accidentally as part of
the https://android-review.googlesource.com/c/1885311/ cleanup. This `-`
didn't refer to reading from stdin, but instead that we wanted to
discard any reject files. On Darwin, the version of `patch` that gets
used can't handle the input file specified at the end of the command
line (as `-i` now gets interpreted as the reject filename). Adding back
the `-` parameter allows things to work again.

This change also adds a bit more output logging around the patch files
so that we can see which patches might need some assistance in the
future (due to increasing differences from the sources).

Test: ./build.py on darwin and look at logs
Change-Id: Id5c868ad87f4925b2ed037bbc81def9b2d4e43e9