Cherry-pick https://reviews.llvm.org/D104556
Bug: http://b/206980172
Bug: http://b/201432564

Cherry-pick: [InstrProfiling] Make CountersPtr in __profd_ relative

This change updates the raw coverage profile format.  It's an
all-or-nothing patch to LLVM and compiler-rt.  Since we use
rust/src/llvm-project to do instrumentation for Rust and the C++
toolchain's compiler-rt, the mismatch breaks coverage.  Apply this
change to Rust toolchain to unbreak coverage.

Test: Test coverage report for android_logger_device_test_src_lib with
clang-r437112 and rust toolchain with this patch.

Change-Id: I53f459d802f13fcf0ba7f050c6764cf3f3c3e006